Reverts changes around items and itemset and adds support for items in RankControl
